11.1.5 добавляет окна сообщений об отправке (успешная/не успешная)
